refactor kwai tests
- moved integration test to it's own file
- created a new integration test that makes sure it returns data even if
it fails
